Fix compilation errors on some non-standard platforms

+++ py/wrapper/yadeWrapper.cpp
@@ -319,7 +319,7 @@ public:
 			        ("Error: Sum of amounts " + boost::lexical_cast<string>(checkSum) + " should not be bigger than 1.0!").c_str());
 			py::throw_error_already_set();
 		}
-		if (py::len(ctList) != (unsigned)amounts.size()) { //avoid unsigned comparison warning
+		if (static_cast<size_t>(py::len(ctList)) != static_cast<size_t>(amounts.size())) { //avoid unsigned comparison warning
 			PyErr_SetString(
 			        PyExc_ValueError,
 			        ("Error: Length of amounts list (" + boost::lexical_cast<string>(amounts.size()) + ") differs from length of template list ("
@@ -1024,10 +1024,10 @@ public:
 		// timespec t1,t2; t1.tv_sec=0; t1.tv_nsec=40000000; /* 40 ms */
 		// while(!OMEGA.isRunning()) nanosleep(&t1,&t2); // wait till we start, so that calling wait() immediately afterwards doesn't return immediately
 		LOG_DEBUG(
-		        "RUN" << ((scene->stopAtIter - scene->iter) > 0
-		                          ? string(" (" + boost::lexical_cast<string>(scene->stopAtIter - scene->iter) + " to go)")
-		                          : string(""))
-		              << "!");
+		        "RUN"
+		        << ((scene->stopAtIter - scene->iter) > 0 ? string(" (" + boost::lexical_cast<string>(scene->stopAtIter - scene->iter) + " to go)")
+		                                                  : string(""))
+		        << "!");
 		if (doWait)
 			wait();
 	}
@@ -1815,4 +1815,3 @@ try {
 	boost::python::handle_exception();
 	throw;
 }
-
